requirements.txt
langchain
langchain_community
langchain_huggingface
faiss-cpu
pymupdf
python-dotenv
streamlit